requirements.txt
ray[serve]
fastapi
uvicorn
httpx
Pillow
PyMuPDF
pyyaml
openai
python-dotenv
numpy